requirements.txt
coverage==4.5.1
gitdb2==2.0.4
GitPython==2.1.11
PyYAML==3.12
smmap2==2.0.4
